+++ packages/web3-providers-ws/test/integration/web_socket_provider_integration.test.ts
@@ -22,13 +22,15 @@ import {
 	JsonRpcSubscriptionResult,
 	JsonRpcId,
 	JsonRpcResponse,
+	ProviderRpcError,
+	Web3ProviderEventCallback,
+	SocketRequestItem,
 } from 'web3-types';
 import { Web3DeferredPromise } from 'web3-utils';
 
 import { Web3WSProviderError } from 'web3-errors';
 import WebSocketProvider from '../../src/index';
-import { OnCloseEvent, WSRequestItem } from '../../src/types';
-import { waitForOpenConnection } from '../fixtures/helpers';
+import { waitForCloseConnection, waitForOpenConnection } from '../fixtures/helpers';
 
 import {
 	getSystemTestProvider,
@@ -45,7 +47,6 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 	let webSocketProvider: WebSocketProvider;
 	let jsonRpcPayload: Web3APIPayload<EthExecutionAPI, 'eth_getBalance'>;
 	// helper function
-	let currentAttempt = 0;
 
 	beforeAll(async () => {
 		clientWsUrl = getSystemTestProvider();
@@ -63,20 +64,18 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 			{},
 			{ delay: 1, autoReconnect: false, maxAttempts: 1 },
 		);
-		currentAttempt = 0;
 	});
 	afterEach(async () => {
 		// make sure we try to close the connection after it is established
 		if (webSocketProvider.getStatus() === 'connecting') {
-			await waitForOpenConnection(webSocketProvider, currentAttempt);
+			await waitForOpenConnection(webSocketProvider);
 		}
-
 		webSocketProvider.disconnect(1000);
 	});
 
 	describe('websocker provider tests', () => {
 		it('should connect', async () => {
-			await waitForOpenConnection(webSocketProvider, currentAttempt);
+			await waitForOpenConnection(webSocketProvider);
 			expect(webSocketProvider).toBeInstanceOf(WebSocketProvider);
 			expect(webSocketProvider.getStatus()).toBe('connected');
 		});
@@ -94,6 +93,9 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 						if (error) {
 							throw new Error(error.message);
 						}
+						if (result?.id !== jsonRpcPayload.id) {
+							return;
+						}
 						expect(result?.id).toBe(jsonRpcPayload.id);
 						resolve();
 					},
@@ -104,6 +106,7 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 		});
 
 		it('should subscribe to `error` event that could happen at the underlying WebSocket connection', async () => {
+			await waitForOpenConnection(webSocketProvider);
 			const errorMsg = 'Custom WebSocket error occurred';
 
 			const errorPromise = new Promise((resolve: Resolve) => {
@@ -113,7 +116,7 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 				});
 			});
 
-			webSocketProvider['_webSocketConnection']?.emit(
+			webSocketProvider['_socketConnection']?.emit(
 				'error',
 				new Web3WSProviderError(errorMsg),
 			);
@@ -122,27 +125,23 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 
 		it('should subscribe to `connect` event', async () => {
 			const openPromise = new Promise((resolve: Resolve) => {
-				webSocketProvider.on('open', () => {
+				webSocketProvider.on('connect', () => {
 					resolve('resolved');
 				});
 			});
 			await expect(openPromise).resolves.toBe('resolved');
 		});
 
-		it('should subscribe to `close` event', async () => {
+		it('should subscribe to `disconnect` event', async () => {
 			const code = 1000;
 
-			const closePromise = new Promise((resolve: Resolve) => {
-				webSocketProvider.on('close', (err?: Error, event?: OnCloseEvent) => {
-					if (err) {
-						throw new Error(err.message);
-					}
-					expect(event?.code).toEqual(code);
-					resolve();
-				});
+			const closePromise = new Promise<ProviderRpcError>(resolve => {
+				webSocketProvider.on('disconnect', ((error: ProviderRpcError) => {
+					expect(error?.code).toEqual(code);
+					resolve(error);
+				}) as Web3ProviderEventCallback<ProviderRpcError>);
 			});
-			currentAttempt = 0;
-			await waitForOpenConnection(webSocketProvider, currentAttempt);
+			await waitForOpenConnection(webSocketProvider);
 			webSocketProvider.disconnect(code);
 			await closePromise;
 		});
@@ -150,20 +149,22 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 
 	describe('disconnect and reset test', () => {
 		it('should disconnect', async () => {
+			await waitForOpenConnection(webSocketProvider);
 			const provider = new WebSocketProvider(
 				clientWsUrl,
 				{},
 				{ delay: 1, autoReconnect: false, maxAttempts: 1 },
 			);
-			await waitForOpenConnection(provider, currentAttempt);
+			await waitForOpenConnection(provider);
 			provider.disconnect(1000);
-			await waitForOpenConnection(provider, currentAttempt, 'disconnected');
+			await waitForCloseConnection(provider);
 			expect(provider.getStatus()).toBe('disconnected');
 		});
 
 		it('should reset', async () => {
+			await waitForOpenConnection(webSocketProvider);
 			class TestReset extends WebSocketProvider {
-				public pendigRequestsSize() {
+				public pendingRequestsSize() {
 					return this._pendingRequestsQueue.size;
 				}
 
@@ -171,11 +172,11 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 					return this._pendingRequestsQueue.size;
 				}
 
-				public setPendingRequest(id: JsonRpcId, reqItem: WSRequestItem<any, any, any>) {
+				public setPendingRequest(id: JsonRpcId, reqItem: SocketRequestItem<any, any, any>) {
 					this._pendingRequestsQueue.set(id, reqItem);
 				}
 
-				public setSentRequest(id: JsonRpcId, reqItem: WSRequestItem<any, any, any>) {
+				public setSentRequest(id: JsonRpcId, reqItem: SocketRequestItem<any, any, any>) {
 					this._sentRequestsQueue.set(id, reqItem);
 				}
 			}
@@ -185,23 +186,23 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 				{ delay: 1, autoReconnect: false, maxAttempts: 1 },
 			);
 
-			await waitForOpenConnection(testResetProvider, currentAttempt);
+			await waitForOpenConnection(testResetProvider);
 
 			const defPromise = new Web3DeferredPromise<JsonRpcResponse<ResponseType>>();
 
-			const reqItem: WSRequestItem<any, any, any> = {
+			const reqItem: SocketRequestItem<any, any, any> = {
 				payload: jsonRpcPayload,
 				deferredPromise: defPromise,
 			};
 
 			testResetProvider.setPendingRequest(jsonRpcPayload.id, reqItem);
-			expect(testResetProvider.pendigRequestsSize()).toBe(1);
+			expect(testResetProvider.pendingRequestsSize()).toBe(1);
 
 			testResetProvider.setSentRequest(jsonRpcPayload.id, reqItem);
 			expect(testResetProvider.sentRequestsSize()).toBe(1);
 
 			testResetProvider.reset();
-			expect(testResetProvider.pendigRequestsSize()).toBe(0);
+			expect(testResetProvider.pendingRequestsSize()).toBe(0);
 			expect(testResetProvider.sentRequestsSize()).toBe(0);
 
 			testResetProvider.disconnect(1000);
@@ -209,17 +210,19 @@ describeIf(isWs)('WebSocketProvider - implemented methods', () => {
 	});
 
 	describe('getStatus get and validate all status tests', () => {
-		it('should getStatus `connecting`', () => {
+		it('should getStatus `connecting`', async () => {
 			expect(webSocketProvider.getStatus()).toBe('connecting');
+			await waitForOpenConnection(webSocketProvider);
 		});
 
 		it('should getStatus `connected`', async () => {
-			await waitForOpenConnection(webSocketProvider, currentAttempt);
+			await waitForOpenConnection(webSocketProvider);
 			expect(webSocketProvider.getStatus()).toBe('connected');
 		});
 		it('should getStatus `disconnected`', async () => {
-			await waitForOpenConnection(webSocketProvider, currentAttempt);
+			await waitForOpenConnection(webSocketProvider);
 			webSocketProvider.disconnect();
+			await waitForCloseConnection(webSocketProvider);
 			expect(webSocketProvider.getStatus()).toBe('disconnected');
 		});
 	});