+++ packages/web3-utils/src/index.ts
@@ -28,3 +28,5 @@ export * as jsonRpc from './json_rpc';
 export * from './web3_deferred_promise';
 export * from './chunk_response_parser';
 export * from './uuid';
+export * from './web3_eip1193_provider';
+export * from './socket_provider';