+++ packages/web3-errors/CHANGELOG.md
@@ -60,3 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Updated Web3.js dependencies (#5664)
 
 ## [Unreleased]
+
+### Changed
+
+-   `main` and `files` entries in `package.json` changed to `lib/` directory from `dist/` (#5739)