+++ packages/web3-providers-ws/CHANGELOG.md
@@ -48,3 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Updated dependencies (#5725)
 
 ## [Unreleased]
+
+### Changed
+
+-   `main` and `files` entries in `package.json` changed to `lib/` directory from `dist/` (#5739)
+-   Refactor to use common SocketProvider class (#5683)